[f] train and test moved to main
